Add Typescript Version Check
**Bug**
With NTVS 1.3, we will depend on TypeScript being installed. With VS15, this is enforced during instalation, but this dependency is not enforced in VS2015. However, we see the majority of users do have TypeScript installed.

**Fix**
Add runtime check to alert users if TypeScript is not installed.

Closes #1259
