+++ Nodejs/Product/Nodejs/Resources.resx
@@ -630,4 +630,7 @@ You will need to restart Visual Studio after installation.</value>
   <data name="CouldNotGetTypeScriptLanguagePreferences" xml:space="preserve">
     <value>Could not retrieve Typescript language preferences. NTVS is not able to load. Please ensure Typescript is properly installed.</value>
   </data>
+  <data name="TypeScriptMinVersionNotInstalled" xml:space="preserve">
+    <value>Node.js Tools requires TypeScript for Visual Studio {0} or higher. Please ensure TypeScript is installed</value>
+  </data>
 </root>
